% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_make_tables.R
\name{f_make_tables}
\alias{f_make_tables}
\title{Create Download Tables with Buttons}
\usage{
f_make_tables(
  data,
  title,
  footnotes = NA,
  data_style = ns_comma,
  data_dir = here("outputs/figdata")
)
}
\description{
Creates CSV and Excel with embedded buttons
}
