% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_excel_tables.R
\name{f_single_excel}
\alias{f_single_excel}
\title{Create Single Excel Table}
\usage{
f_single_excel(
  title,
  info,
  notes = c(NA),
  df,
  sheet,
  tablename,
  num_cols = NA,
  pct_cols = NA
)
}
\description{
Creates formatted Excel table
}
