% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_break_axis.R
\name{fn_break_axis}
\alias{fn_break_axis}
\title{Add Broken Axis Symbol to Plotly Chart}
\usage{
fn_break_axis(p, linecolor, ref_line = list())
}
\description{
Adds broken axis symbol to plotly charts
This function will insert a broken axis symbol on the y axis when called.
It takes three arguments:
1. p - a plotly object
2. linecolor - a string containing colour code of axis, you should match the
   colour defined here to that of the axis lines
3. ref_line - optional argument, list definition of reference line
}
